Handle disconnected  popup in passwordless login.
The passwordless login method sends the user a link in an email.  When the
user follows this link a new browser window opens.

However the original popup window will not redirect and remains open
with a page allowing to resend the link.

This change has the popups parent window listens whether an
authentication is established. For this an 'authenticated' event
was introduced. The 'authenticated' event is emitted in response to a
local storage 'anvil.connect' event when the user is authenticated.
This reacts to an authentiation performed in another window or tab.

For the eventing a new dependency was added to `tiny-emitter`.
Here are a few alternative event emitters:
  - https://github.com/Olical/EventEmitter
    seems very elaborate, a bit too much weight for our scenario
  - https://github.com/component/emitter
    a similar alternative to tiny-emitter. More or less a coin-toss.
  - https://github.com/jeromeetienne/microevent.js
    seemed too microi (perhaps)!
  - https://github.com/joaquimserafim/tiny-eventemitter.
    not too popular. Might be an alternative too.

Presumably this could also be handled with the server. In this case the
'Sent mail' page would have to monitor whether server whether the
session is established. However one would want to avoid publishing the
email link to that window as one could easily steal it and allowing
anyone to login. Perhaps divulging the secret could be avoided similar to
how OpenID Connect session management works.

Another disadvantage would be that the server would have to be either
polled or sending server events. Again it might be possible to do this
with the existing session events.

An advantage would be that this would allow the normal OpenID connect
authentication flow to happen.

One issue with this solution is that the page opened by the emailed link
activation may not self close as some browser may not allow this from a
script in this case. I did see this work in Chrome and Safari but not
Firefox.