Add support for better logging.
During development for sure there is some need to log how the client
interacts with connect.

While one might want a capability to submit client errors to a server,
this is not what this is for.

Instead this allows to have essentially console statement which are can
be enabled optionally by setting

  `localStorage.debug = true`

in the browser dev tools console. To stop logging use

  `delete localStorage.debug`

The logging library used is https://github.com/latentflip/bows.

I have read about https://github.com/visionmedia/debug which seems more
popular but it was mentioned that it does not link to the code in its
messages. This is not helpful in my opinion.

I played around a bit with bows and it pretty good to me.