Anvil.nonce and Anvil.sha256url to use web crypto.
Both Anvil.nonce and Anvil.sha256url now return promises.

Unfortunately this means that all consumers will have to adjust.
However apparently there is no real way around this fact.

This required implementing the following conversion or encodings for
ArrayBuffers:
  1. encode/decode str as utf-8
  2. convert buffer to base64url string
  3. in the context of seriale base64 conversion to/from buffer
     was needed.

Implementing utf-8 support should be relatively straightforward with
ES6. However I decided to use the npm:text-encoder-lite library instead
which is mentioned by MDN (https://developer.mozilla.org). It appears
that at the moment there is not yet widespread support in the browsers
for TextEncode/TextDecode and so this seemed like a pretty good
approach. However it might also be on option to piggy back on node
packages when using browserify or jspm. However I am not sure this is
desirable.

To install this library I had to shim it this should perhaps be
submitted to the jspm registry. See package.json for the override.