Split off extended test useful to gen tokens
The extended tests are not required for verifying jws tokens.
However in case one needs to generate new tokens the code can be useful.
To observe the bows log statement one must set debug=true in
localStorage.

The extended tests can be run as follows:
  1. npm run karma-browsers-extended
  2. npm run karma-run

Currently the extended test is known to fail on [Safari 9.0.2 (Mac OS X
10.10.5)] and passes on
- [Firefox 43.0.0 (Mac OS X 10.10.0)]
- [Chrome 47.0.2526 (Mac OS X 10.10.5)]